fix the tests, refine the liquid file handling and add a unit-like test to describe the special handling of liquid files

+++ src/copy.rs
@@ -1,12 +1,14 @@
-use anyhow::{Ok, Result};
+use anyhow::{bail, Ok, Result};
 use console::style;
-use log::warn;
+use log::{debug, warn};
 use std::{
-    fs::{copy, read_dir, remove_file, DirEntry},
+    fs::{copy, read_dir, remove_file},
     path::Path,
 };
 
-pub(crate) fn copy_files_recursively(
+const LIQUID_SUFFIX: &str = ".liquid";
+
+pub fn copy_files_recursively(
     src: impl AsRef<Path>,
     dst: impl AsRef<Path>,
     overwrite: bool,
@@ -17,8 +19,6 @@ pub(crate) fn copy_files_recursively(
         let src_entry = src_entry?;
         let filename = src_entry.file_name().to_string_lossy().to_string();
         let entry_type = src_entry.file_type()?;
-
-        println!("[debug] src = {src_entry:?} dst = {dst_path:?}");
 
         if entry_type.is_dir() {
             // we skip the .git directory
@@ -31,7 +31,7 @@ pub(crate) fn copy_files_recursively(
             }
             copy_files_recursively(src_entry.path(), dst_dir, overwrite)?;
         } else if entry_type.is_file() {
-            copy_file(src_entry, dst_path, overwrite)?;
+            copy_file(&src_entry.path(), dst_path, overwrite)?;
         } else {
             // todo: maybe we better emit a warning but continue processing the other files
             warn!(
@@ -50,15 +50,30 @@ pub(crate) fn copy_files_recursively(
 
 /// move a file from src to dst, possibly overwriting existing files if overwrite is true
 /// if the file has a .liquid suffix, the suffix will be removed in the destination, and overwritten if existing
-fn copy_file(src_entry: DirEntry, dst: &Path, overwrite: bool) -> Result<()> {
-    let filename = src_entry.file_name().to_string_lossy().to_string();
+fn copy_file(src_path: &Path, dst: &Path, overwrite: bool) -> Result<()> {
+    let filename = src_path.file_name().unwrap().to_string_lossy().to_string();
     let dst_path = dst.join(&filename);
-    let src_path = src_entry.path();
+    let mut overwrite = overwrite;
 
-    if let Some(new_filename) = filename.strip_suffix(".liquid") {
+    if let Some(new_filename) = filename.strip_suffix(LIQUID_SUFFIX) {
+        if src_path.with_file_name(new_filename).exists() {
+            // if there is a file without the .liquid suffix, we want to set overwrite to true
+            // so that this liquid file takes precedence over the existing file
+            debug!("A non-liquid file exists for {filename}, overwriting it with the liquid file");
+            overwrite = true;
+        }
+
         // move the file to a new filename without the .liquid suffix, in any case
         let dst_path = dst.join(new_filename);
-        safe_copy(&src_path, &dst_path, overwrite)?;
+        safe_copy(src_path, &dst_path, overwrite)?;
+    } else if src_path
+        .with_file_name(format!("{filename}{LIQUID_SUFFIX}"))
+        .exists()
+    {
+        // there is a liquid file for this non-liquid file,
+        // so we skip this one, so that the liquid file takes precedence
+        debug!("A liquid file exists for {filename}, skipping the non-liquid file");
+        return Ok(());
     } else {
         // if the file doesn't have a .liquid suffix, just copy it
         // possibly overwriting existing files if overwrite is true
@@ -69,6 +84,16 @@ fn copy_file(src_entry: DirEntry, dst: &Path, overwrite: bool) -> Result<()> {
 }
 
 fn safe_copy(src_path: &Path, dst_path: &Path, overwrite: bool) -> Result<()> {
+    if dst_path.exists() && !overwrite {
+        bail!(
+            "{} {} `{}` {}",
+            crate::emoji::ERROR,
+            style("File already exists").bold().red(),
+            style(dst_path.display()).bold(),
+            style("and `--overwrite` was not passed")
+        )
+    }
+
     if dst_path.exists() && overwrite {
         remove_file(dst_path)?;
         copy(src_path, dst_path)?;
@@ -77,4 +102,71 @@ fn safe_copy(src_path: &Path, dst_path: &Path, overwrite: bool) -> Result<()> {
     }
 
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use tempfile::tempdir;
+
+    #[test]
+    fn test_overwriting_behavior() {
+        let tmp = tempdir().unwrap();
+        let f1 = tmp.path().join("README.md");
+        std::fs::write(&f1, "A README").unwrap();
+        let f2 = tmp.path().join("README.md.liquid");
+        std::fs::write(&f2, "A README liquid file").unwrap();
+
+        assert!(
+            safe_copy(f1.as_path(), f2.as_path(), false).is_err(),
+            "we do not allow overwriting without the flag set"
+        );
+        assert!(
+            safe_copy(f1.as_path(), f2.as_path(), true).is_ok(),
+            "we do allow overwriting with the flag set"
+        );
+        assert_eq!(std::fs::read_to_string(f2.as_path()).unwrap(), "A README");
+    }
+
+    #[test]
+    fn test_special_liquid_file_handling() {
+        let tmp = tempdir().unwrap();
+        let f1 = tmp.path().join("README.md");
+        std::fs::write(&f1, "A README").unwrap();
+        let f2 = tmp.path().join("README.md.liquid");
+        std::fs::write(&f2, "A README liquid file").unwrap();
+
+        let tmp2 = tempdir().unwrap();
+
+        // copy the non liquid file first, should not copy anything
+        copy_file(f1.as_path(), tmp2.path(), false).unwrap();
+        assert!(
+            !tmp2.path().join("README.md").exists(),
+            "the file should not be copied"
+        );
+
+        // copy the liquid file, should copy the liquid file and remove the .liquid suffix
+        copy_file(f2.as_path(), tmp2.path(), false).unwrap();
+        assert!(
+            tmp2.path().join("README.md").exists(),
+            "the file should be copied and the .liquid suffix removed"
+        );
+        assert_eq!(
+            std::fs::read_to_string(tmp2.path().join("README.md")).unwrap(),
+            "A README liquid file"
+        );
+
+        // copy the liquid file, while the same file without the .liquid suffix exists,
+        // should overwrite the existing file, as if the overwrite flag was set
+        let f4 = tmp2.path().join("README.md");
+        std::fs::write(&f4, "Existing file, should be overwritten").unwrap();
+        assert!(
+            copy_file(f2.as_path(), tmp2.path(), false).is_ok(),
+            "the file should be copied"
+        );
+        assert_eq!(
+            std::fs::read_to_string(tmp2.path().join("README.md")).unwrap(),
+            "A README liquid file"
+        );
+    }
 }
